+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add a view to cancel reminder mails, link added in the footer of mails
 - Prevent multiple join to BBB meetings
 - Introduce anonymous_id mechanism
+- Create a direct access to a video out of an LMS context for a public 
+  or LTI ressource, based on a liveregistration
 
 ### Fixed